v2.6:把状态队列从报文状态码提取改为从共享内存提取，并把状态队列修改为u64类型

+++ aflnet.c
@@ -1802,7 +1802,7 @@ int parse_net_config(u8* net_config, u8* protocol, u8** ip_address, u32* port)
   return 0;
 }
 
-u8* state_sequence_to_string(unsigned int *stateSequence, unsigned int stateCount) {
+u8* state_sequence_to_string(u64 *stateSequence, unsigned int stateCount) {
   u32 i = 0;
 
   u8 *out = NULL;
