+++ hash.h
@@ -40,7 +40,7 @@
 
 #define ROL64(_x, _r)  ((((u64)(_x)) << (_r)) | (((u64)(_x)) >> (64 - (_r))))
 
-static inline u32 hash32(const void* key, u32 len, u32 seed) {
+static inline u64 hash32(const void* key, u32 len, u32 seed) {
 
   const u64* data = (u64*)key;
   u64 h1 = seed ^ len;
